requirements.txt

DrissionPage>=4.0.5.7
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.25.0
Pillow>=9.0.0
customtkinter>=5.2.0